feat(core): change selection to tab and remove display of type and tooltip. (#8805)
* feat(core): replace select with tabs for schema selection.

* feat(core): change selection to tab and remove display of type and tooltip.

* fix: remove obsolete component and add class.

* chore(core): task enclosed under a border.

* fixes enclosed border only for anyOf type

---------
